feat(common.sh): add function get_os_distro to determine OS distribution codename for better compatibility feat(headscale.sh, tailscale.sh): add logic to determine OS distribution codename using lsb_release or cat command for improved reliability feat(apt.sh): add function install_os_specific_packages to install OS-specific packages for Debian based on OS distribution codename

+++ dot_config/bin/installers/tailscale.sh
@@ -6,14 +6,19 @@ source "$(dirname "$BASH_SOURCE")/../init/init.sh"
 
 # Function to install Tailscale on Linux
 install_tailscale_linux() {
-    if command_exists curl && command_exists lsb_release && command_exists sudo; then
+    if command_exists curl && (command_exists lsb_release || command_exists cat); then
         echo_with_color "$GREEN_COLOR" "Installing Tailscale..."
         local RELEASE
         local DISTRO
-        RELEASE=$(lsb_release -cs)
-        DISTRO=$(lsb_release -is | tr '[:upper:]' '[:lower:]')
+        if command_exists lsb_release; then
+            RELEASE=$(lsb_release -cs)
+            DISTRO=$(lsb_release -is | tr '[:upper:]' '[:lower:]')
+        elif command_exists cat; then
+            RELEASE=$(cat /etc/os-release | grep -oP '(?<=VERSION_CODENAME=).+' | tr -d '"')
+            DISTRO=$(grep -oP '(?<=^ID=).+' /etc/os-release | tr -d '"')
+        fi
         if [ -z "$RELEASE" ]; then
-            exit_with_error "Could not determine the distribution codename with lsb_release."
+            exit_with_error "Could not determine the distribution codename. Exiting."
         fi
 
         # Add the Tailscale repository signing key and repository
@@ -21,13 +26,21 @@ install_tailscale_linux() {
         curl -fsSL "https://pkgs.tailscale.com/stable/${DISTRO}/${RELEASE}.tailscale-keyring.list" | sudo tee /etc/apt/sources.list.d/tailscale.list
 
         # Update the package list and install Tailscale
-        sudo apt-get update || exit_with_error "Failed to update package list. Exiting."
-        sudo apt-get install tailscale -y || exit_with_error "Failed to install Tailscale. Exiting."
+        if command_exists apt-get; then
+            sudo apt-get update || exit_with_error "Failed to update package list. Exiting."
+            sudo apt-get install tailscale -y || exit_with_error "Failed to install Tailscale. Exiting."
+        elif command_exists dnf; then
+            sudo dnf install -y tailscale || exit_with_error "Failed to install Tailscale. Exiting."
+        elif command_exists zypper; then
+            sudo zypper install -y tailscale || exit_with_error "Failed to install Tailscale. Exiting."
+        else
+            exit_with_error "Package manager not found. Please install Tailscale manually."
+        fi
 
         # Start Tailscale and authenticate
         authenticate_tailscale
     else
-        exit_with_error "Required command(s) are missing. Please ensure curl, lsb_release, and sudo are installed to proceed."
+        exit_with_error "Required command(s) are missing. Please ensure curl and either lsb_release or cat are installed to proceed."
     fi
 }
